+++ plonky2x/src/frontend/eth/rlp/mod.rs
@@ -1,1 +1,2 @@
 pub mod builder;
+pub mod generators;
